refactor: model package
