+++ src/main/java/blackjack/model/Card.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import java.util.Objects;
