+++ src/main/java/blackjack/model/CardScore.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 public class CardScore {
