+++ src/main/java/blackjack/model/CardType.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 public enum CardType {
     DIAMOND("다이아몬드"), HEART("하트"), SPADE("스페이드"), CLOVER("클로버");
