+++ src/main/java/blackjack/model/Player.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 public interface Player {
     String getName();
