+++ src/main/java/blackjack/model/RandomShuffleStrategy.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import java.util.Collections;
 import java.util.Stack;
