+++ src/main/java/blackjack/model/ShuffleStrategy.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 import java.util.Stack;
