+++ src/test/java/blackjack/model/CardTest.java
@@ -1,5 +1,8 @@
-package blackjack;
+package blackjack.model;
 
+import blackjack.model.Card;
+import blackjack.model.CardType;
+import blackjack.model.CardValue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
