+++ src/test/java/blackjack/model/GamerTest.java
@@ -1,5 +1,6 @@
-package blackjack;
+package blackjack.model;
 
+import blackjack.model.*;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
